[refactor] Divide navigator component
